Re-enable this code in a way where it now hopefully works.
After speaking with Kenneth Nielsen, it turns out that this is the
code that handles in-line replacements of magic labels. It turns
out that also included some of the meta-data features of the data
pages, so it was necessecary to fix it :-)

Hopefully it now works as it has always done. One day it should be
re-written in more general way.